+++ doctests.hs
@@ -1,4 +1,7 @@
-import Test.DocTest
+import System.Exit (exitWith)
+import System.Process (system)
 
 main :: IO ()
-main = doctest ["-isrc", "src"]
+main = do
+  exitWith =<< system "cabal repl --with-ghc=doctest"
+  -- See README of doctest https://hackage.haskell.org/package/doctest