+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# Unreleased 0.12.0
+
+* Pull updates in cryptostore, fixing build with crypton.
+* Stabilize ordering of keys, both in the header and in claims, resolving
+  [#2](https://github.com/puffnfresh/haskell-jwt/issues/2).
+* Fix and reenable doctests.
+* Drop support for aeson < 2.
+
 # 2021-12-11 0.11.0
 
 * Added support for RSA256 Public Key verification. This in turn means that some